# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ISP Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /customers:
    get:
      operationId: listCustomers
      tags: [customers]
      summary: List all customers
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of customers
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerList"
    post:
      operationId: createCustomer
      tags: [customers]
      summary: Create a new customer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
  /customers/{id}:
    get:
      operationId: getCustomer
      tags: [customers]
      summary: Get customer by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Customer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
        "404":
          description: Not found
    patch:
      operationId: updateCustomer
      tags: [customers]
      summary: Update a customer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerUpdate"
      responses:
        "200":
          description: Updated customer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
    delete:
      operationId: deleteCustomer
      tags: [customers]
      summary: Delete a customer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── PLANS ─────────────────────────────────────────────────────────────────────
...
  /invoices:
    get:
      operationId: listInvoices
      tags: [invoices]
      summary: List all invoices
      parameters:
        - name: customerId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Invoice list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvoiceList"
    post:
      operationId: createInvoice
      tags: [invoices]
      summary: Create an invoice
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InvoiceInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
  /invoices/{id}:
    get:
      operationId: getInvoice
      tags: [invoices]
      summary: Get invoice by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Invoice
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
    patch:
      operationId: updateInvoice
      tags: [invoices]
      summary: Update an invoice
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InvoiceUpdate"
      responses:
        "200":
          description: Updated invoice
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
    delete:
      operationId: deleteInvoice
      tags: [invoices]
      summary: Delete an invoice
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── PAYMENTS ──────────────────────────────────────────────────────────────────
...
  /equipment:
    get:
      operationId: listEquipment
      tags: [equipment]
      summary: List all network equipment
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: type
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Equipment list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Equipment"
    post:
      operationId: createEquipment
      tags: [equipment]
      summary: Add network equipment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EquipmentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Equipment"
  /equipment/{id}:
    get:
      operationId: getEquipment
      tags: [equipment]
      summary: Get equipment by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Equipment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Equipment"
    patch:
      operationId: updateEquipment
      tags: [equipment]
      summary: Update equipment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EquipmentUpdate"
      responses:
        "200":
          description: Updated equipment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Equipment"
    delete:
      operationId: deleteEquipment
      tags: [equipment]
      summary: Delete equipment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── IP POOLS ──────────────────────────────────────────────────────────────────
  /ip-pools:
    get:
      operationId: listIpPools
      tags: [ip-pools]
      summary: List all IP pools
      responses:
        "200":
          description: IP pools
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/IpPool"
    post:
      operationId: createIpPool
      tags: [ip-pools]
      summary: Create an IP pool
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IpPoolInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IpPool"
  /ip-pools/{id}:
    get:
      operationId: getIpPool
      tags: [ip-pools]
      summary: Get IP pool by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: IP pool
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IpPool"
    patch:
      operationId: updateIpPool
      tags: [ip-pools]
      summary: Update an IP pool
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IpPoolUpdate"
      responses:
        "200":
          description: Updated IP pool
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IpPool"
    delete:
      operationId: deleteIpPool
      tags: [ip-pools]
      summary: Delete an IP pool
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
components:
  schemas:
...
    Customer:
      type: object
      required:
        - id
        - name
        - email
        - phone
        - address
        - status
        - createdAt
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        address:
          type: string
        status:
          type: string
          enum: [active, suspended, terminated]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
    CustomerList:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Customer"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    CustomerInput:
      type: object
      required: [name, email, phone, address]
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        address:
          type: string
        status:
          type: string
          enum: [active, suspended, terminated]
        notes:
          type: string
    CustomerUpdate:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        address:
          type: string
        status:
          type: string
          enum: [active, suspended, terminated]
        notes:
          type: ["string", "null"]

    # ── PLANS ──────────────────────────────────────────────────────────────────
...
    Invoice:
      type: object
      required:
        - id
        - customerId
        - amount
        - status
        - dueDate
        - createdAt
      properties:
        id:
          type: integer
        customerId:
          type: integer
        subscriptionId:
          type: ["integer", "null"]
        amount:
          type: number
        tax:
          type: ["number", "null"]
        total:
          type: number
        status:
          type: string
          enum: [draft, sent, paid, overdue, cancelled]
        dueDate:
          type: string
        paidAt:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
        customer:
          $ref: "#/components/schemas/Customer"
    InvoiceList:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Invoice"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    InvoiceInput:
      type: object
      required: [customerId, amount, dueDate]
      properties:
        customerId:
          type: integer
        subscriptionId:
          type: integer
        amount:
          type: number
        tax:
          type: number
        status:
          type: string
          enum: [draft, sent, paid, overdue, cancelled]
        dueDate:
          type: string
        notes:
          type: string
    InvoiceUpdate:
      type: object
      properties:
        amount:
          type: number
        tax:
          type: ["number", "null"]
        status:
          type: string
          enum: [draft, sent, paid, overdue, cancelled]
        dueDate:
          type: string
        paidAt:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]

    # ── PAYMENTS ───────────────────────────────────────────────────────────────
...
    Equipment:
      type: object
      required:
        - id
        - name
        - type
        - model
        - ipAddress
        - status
        - createdAt
      properties:
        id:
          type: integer
        name:
          type: string
        type:
          type: string
          enum: [router, switch, olt, onu, access_point, server, other]
        model:
          type: string
        brand:
          type: ["string", "null"]
        ipAddress:
          type: string
        macAddress:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        status:
          type: string
          enum: [online, offline, maintenance, unknown]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
    EquipmentInput:
      type: object
      required: [name, type, model, ipAddress]
      properties:
        name:
          type: string
        type:
          type: string
          enum: [router, switch, olt, onu, access_point, server, other]
        model:
          type: string
        brand:
          type: string
        ipAddress:
          type: string
        macAddress:
          type: string
        location:
          type: string
        status:
          type: string
          enum: [online, offline, maintenance, unknown]
        notes:
          type: string
    EquipmentUpdate:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
          enum: [router, switch, olt, onu, access_point, server, other]
        model:
          type: string
        brand:
          type: ["string", "null"]
        ipAddress:
          type: string
        macAddress:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        status:
          type: string
          enum: [online, offline, maintenance, unknown]
        notes:
          type: ["string", "null"]

    # ── IP POOLS ───────────────────────────────────────────────────────────────
    IpPool:
      type: object
      required:
        - id
        - name
        - network
        - gateway
        - subnetMask
        - totalIps
        - usedIps
        - createdAt
      properties:
        id:
          type: integer
        name:
          type: string
        network:
          type: string
          description: CIDR notation e.g. 192.168.1.0/24
        gateway:
          type: string
        subnetMask:
          type: string
        dns1:
          type: ["string", "null"]
        dns2:
          type: ["string", "null"]
        totalIps:
          type: integer
        usedIps:
          type: integer
        description:
          type: ["string", "null"]
        createdAt:
          type: string
    IpPoolInput:
      type: object
      required: [name, network, gateway, subnetMask]
      properties:
        name:
          type: string
        network:
          type: string
        gateway:
          type: string
        subnetMask:
          type: string
        dns1:
          type: string
        dns2:
          type: string
        description:
          type: string
    IpPoolUpdate:
      type: object
      properties:
        name:
          type: string
        gateway:
          type: string
        subnetMask:
          type: string
        dns1:
          type: ["string", "null"]
        dns2:
          type: ["string", "null"]
        description:
          type: ["string", "null"]