# lib/api-zod/src/generated/types/customerUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ISP Management System API
 * OpenAPI spec version: 0.1.0
 */
import type { CustomerUpdateStatus } from './customerUpdateStatus';

export interface CustomerUpdate {
  name?: string;
  email?: string;
  phone?: string;
  address?: string;
  status?: CustomerUpdateStatus;
  /** @nullable */
  notes?: string | null;
}